LayoutTests/fast/events/touch/ios/long-press-on-image-expected.txt

Verifies that long pressing an image presents a context menu that shows the image name. To manually test, long press on the image.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context menu triggered on image
PASS results.items.includes('Copy') is true
PASS successfullyParsed is true

TEST COMPLETE
